refactor(insurance): move import gate helpers into shared/insurance
The preview and confirmation-gate modules landed in a new generic
src/shared/import/ area, but insurance is their only consumer. Move them
next to record-update-plan.ts / record-update-apply.ts so this change
touches only insurance files, and reword the header comments so they read
accurately from the insurance layer.

Pure move: same exports, same behavior, same tests. The modules stay
command-agnostic, so cml import as-expression-set can still adopt them when
its gate lands.
